Troubleshooting

+++ cPWP/bamsToPWP.cpp
@@ -184,7 +184,7 @@ int convertANGSDcountsToBinary(std::string angsdPrefix, std::string binaryOutput
 
 
 
-int calcPWPfromBinaryFile (std::string binaryFile, int numLoci, const int numIndividuals, std::string outFile) {
+int calcPWPfromBinaryFile (std::string binaryFile, unsigned long long int numLoci, const int numIndividuals, std::string outFile) {
     typedef unsigned char BYTE;
     
     //****MODIFY THIS TO ONLY READ IN N LOCI AT A TIME, INSTEAD OF USING THE ENTIRE FILE****
@@ -210,7 +210,7 @@ int calcPWPfromBinaryFile (std::string binaryFile, int numLoci, const int numInd
         
         // We now have an array of numIndividuals * 2 (major and minor allele) * 1million (loci)
         //int totalLoci = (int)size / (numIndividuals*2); // The 1 million locus file has 999,999 sites in it (because of header line)
-        int totalLoci = numLoci;
+        unsigned long long int totalLoci = numLoci;
         //int totalLoci = size/(272*2);
         std::vector< std::vector<long double> > pwp(numIndividuals, std::vector<long double>(numIndividuals,0));
         std::vector< std::vector<unsigned long long int> > weightings(numIndividuals, std::vector<unsigned long long int>(numIndividuals,0));
